+++ brewery-host/src/main/java/wadosm/breweryhost/logic/brewing/BrewingController.java
@@ -58,7 +58,7 @@ public class BrewingController {
     @Data
     static class CalibrateTemperatureDto {
         private final Integer side;
-        private final Float value;
+        private final Double value;
     }
 
     @PostMapping("/calibrateTemperature")
