+++ brewery-host/src/test/java/wadosm/breweryhost/logic/brewing/AlarmProviderTest.java
@@ -21,43 +21,43 @@ class AlarmProviderTest {
 
     static Stream<Arguments> alarm_disabled_by_default() {
         return Stream.of(
-                Arguments.of((Float) null, BrewingSettings.builder()
+                Arguments.of(null, BrewingSettings.builder()
                         .enabled(true)
                         .temperatureAlarmEnabled(true)
-                        .destinationTemperature(50.0f)
+                        .destinationTemperature(50.0)
                         .build()),
 
-                Arguments.of(49.0f, BrewingSettings.builder()
+                Arguments.of(49.0, BrewingSettings.builder()
                         .enabled(true)
                         .temperatureAlarmEnabled(true)
-                        .destinationTemperature(50.0f)
+                        .destinationTemperature(50.0)
                         .build()),
 
-                Arguments.of(100.0f, BrewingSettings.builder()
+                Arguments.of(100.0, BrewingSettings.builder()
                         .enabled(false)
                         .temperatureAlarmEnabled(true)
-                        .destinationTemperature(50.0f)
+                        .destinationTemperature(50.0)
                         .build()),
 
-                Arguments.of(100.0f, BrewingSettings.builder()
+                Arguments.of(100.0, BrewingSettings.builder()
                         .enabled(true)
                         .temperatureAlarmEnabled(false)
-                        .destinationTemperature(50.0f)
+                        .destinationTemperature(50.0)
                         .build()),
 
-                Arguments.of(100.0f, BrewingSettings.builder()
+                Arguments.of(100.0, BrewingSettings.builder()
                         .enabled(true)
                         .temperatureAlarmEnabled(true)
                         .destinationTemperature(null)
                         .build()),
 
-                Arguments.of(100.0f, BrewingSettings.builder().build())
+                Arguments.of(100.0, BrewingSettings.builder().build())
         );
     }
 
     @ParameterizedTest
     @MethodSource
-    void alarm_disabled_by_default(Float currentTemperature) {
+    void alarm_disabled_by_default(Double currentTemperature) {
         // given
         BreweryInterface breweryInterface = mock(BreweryInterface.class);
 
@@ -82,14 +82,14 @@ class AlarmProviderTest {
         when(settingsProvider.getBrewingSettings()).thenReturn(BrewingSettings.builder()
                 .enabled(true)
                 .temperatureAlarmEnabled(true)
-                .destinationTemperature(50.0f)
+                .destinationTemperature(50.0)
                 .build()
         );
 
         AlarmProvider alarmProvider = new AlarmProvider(breweryInterface, settingsProvider, mockConfigProvider(), mock(TimeProvider.class));
 
         // when
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         // then
         verify(breweryInterface).setAlarm(true);
@@ -110,7 +110,7 @@ class AlarmProviderTest {
         when(settingsProvider.getBrewingSettings()).thenReturn(BrewingSettings.builder()
                 .enabled(true)
                 .temperatureAlarmEnabled(true)
-                .destinationTemperature(50.0f)
+                .destinationTemperature(50.0)
                 .build()
         );
 
@@ -125,16 +125,16 @@ class AlarmProviderTest {
 
         // when
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(1000));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(1500));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(2500));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(3500));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         // then
         ArgumentCaptor<Boolean> alarmSequence = ArgumentCaptor.forClass(Boolean.class);
@@ -152,7 +152,7 @@ class AlarmProviderTest {
         when(settingsProvider.getBrewingSettings()).thenReturn(BrewingSettings.builder()
                 .enabled(true)
                 .temperatureAlarmEnabled(true)
-                .destinationTemperature(50.0f)
+                .destinationTemperature(50.0)
                 .build()
         );
 
@@ -167,19 +167,19 @@ class AlarmProviderTest {
 
         // when
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(1000));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(3500));
-        alarmProvider.handleAlarm(49.0f);
+        alarmProvider.handleAlarm(49.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(4500));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(6400));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         when(timeProvider.getCurrentTime()).thenReturn(Instant.ofEpochMilli(6600));
-        alarmProvider.handleAlarm(51.0f);
+        alarmProvider.handleAlarm(51.0);
 
         // then
         ArgumentCaptor<Boolean> alarmSequence = ArgumentCaptor.forClass(Boolean.class);
